Update SonosAdvPlayer.groovy
Default to "repeat all". If the favorite is a single track, it doesn't change the behavior, as repeating a single track via repeat all or repeat one both result in the same result, the single track repeating over and over. If someone loads a playlist, album, etc, then it'll now repeat that entire favorite, rather than the first track.